+++ cmd/eval-seedgen/main.go
@@ -280,504 +280,6 @@ func generateRelationalSeed(seedID string, worlds int, outDir, suitePath string)
 	return writeJSONL(suitePath, suite)
 }
 
-func generateFeedbackRecallHardcasesV1(outDir, suitePath string) error {
-	generated := buildFeedbackRecallHardcasesV1()
-	corpusCount := 0
-	hardNegativeCount := 0
-	dreamCount := 0
-	for _, item := range generated {
-		corpusCount += len(item.Corpus)
-		hardNegativeCount += len(item.HardNegatives)
-		dreamCount += len(item.Dreams)
-	}
-	manifest := evalharness.SeedManifest{
-		SchemaVersion:     evalharness.SeedSchemaVersion,
-		SeedID:            presetFeedbackRecallHardcases,
-		Description:       "Feedback-derived recall hardcases for exact identifiers, metadata recall text, authority tiers, and dream relevance filtering.",
-		GeneratedAt:       "2026-07-08T00:00:00Z",
-		CorpusFile:        "corpus.jsonl",
-		CasesFile:         "cases.jsonl",
-		QrelsFile:         "qrels.jsonl",
-		AnswersFile:       "answers.jsonl",
-		HardNegativesFile: "hard_negatives.jsonl",
-		TransformsFile:    "transforms.jsonl",
-		DreamsFile:        "expected_dreams.jsonl",
-		LicensesFile:      "licenses.md",
-		Counts: map[string]int{
-			"cases":          len(generated),
-			"corpus":         corpusCount,
-			"docs_per_case":  3,
-			"qrels":          len(generated),
-			"answers":        len(generated),
-			"transforms":     len(generated),
-			"hard_negatives": hardNegativeCount,
-			"dreams":         dreamCount,
-		},
-		Sources: []evalharness.SeedSource{
-			{
-				Name:    "dense-mem synthetic recall feedback hardcases",
-				License: "Project-local synthetic data",
-				Notes:   "Generated by cmd/eval-seedgen from observed recall feedback failure modes; contains no production data.",
-			},
-		},
-	}
-
-	corpus := make([]evalharness.CorpusItem, 0, corpusCount)
-	cases := make([]evalharness.Case, 0, len(generated))
-	qrels := make([]evalharness.QRel, 0, len(generated))
-	answers := make([]evalharness.AnswerLabel, 0, len(generated))
-	hardNegatives := make([]hardNegativeRow, 0, hardNegativeCount)
-	transforms := make([]transformRow, 0, len(generated))
-	dreams := make([]evalharness.ExpectedDream, 0, dreamCount)
-	suite := make([]evalharness.SuiteCase, 0, len(generated))
-	for _, item := range generated {
-		corpus = append(corpus, item.Corpus...)
-		cases = append(cases, item.Case)
-		qrels = append(qrels, item.QRel)
-		answers = append(answers, item.Answer)
-		hardNegatives = append(hardNegatives, item.HardNegatives...)
-		transforms = append(transforms, item.Transform)
-		dreams = append(dreams, item.Dreams...)
-		suite = append(suite, item.SuiteCase)
-	}
-	if err := writeJSONFile(filepath.Join(outDir, "seed_manifest.json"), manifest); err != nil {
-		return err
-	}
-	if err := writeJSONL(filepath.Join(outDir, "corpus.jsonl"), corpus); err != nil {
-		return err
-	}
-	if err := writeJSONL(filepath.Join(outDir, "cases.jsonl"), cases); err != nil {
-		return err
-	}
-	if err := writeJSONL(filepath.Join(outDir, "qrels.jsonl"), qrels); err != nil {
-		return err
-	}
-	if err := writeJSONL(filepath.Join(outDir, "answers.jsonl"), answers); err != nil {
-		return err
-	}
-	if err := writeJSONL(filepath.Join(outDir, "hard_negatives.jsonl"), hardNegatives); err != nil {
-		return err
-	}
-	if err := writeJSONL(filepath.Join(outDir, "transforms.jsonl"), transforms); err != nil {
-		return err
-	}
-	if err := writeJSONL(filepath.Join(outDir, "expected_dreams.jsonl"), dreams); err != nil {
-		return err
-	}
-	if err := writeText(filepath.Join(outDir, "licenses.md"), feedbackHardcaseLicense()); err != nil {
-		return err
-	}
-	return writeJSONL(suitePath, suite)
-}
-
-func buildFeedbackRecallHardcasesV1() []generatedCase {
-	return []generatedCase{
-		buildFeedbackIdentifierMetadataCase(),
-		buildFeedbackToolIdentifierCase(),
-		buildFeedbackPathConfigCase(),
-		buildFeedbackAuthorityFactCase(),
-		buildFeedbackBranchRunCase(),
-		buildFeedbackDreamFilterCase(),
-	}
-}
-
-func buildFeedbackIdentifierMetadataCase() generatedCase {
-	caseID := feedbackCaseID(1)
-	requiredID := caseID + "_required"
-	negative1ID := caseID + "_negative_01"
-	negative2ID := caseID + "_negative_02"
-	query := "What changed in PR #56 commit 2e4b0cc for evalListEdges LIMIT handling?"
-	return feedbackGeneratedCase(
-		caseID,
-		"identifier_metadata",
-		query,
-		"Resolve the PR and commit from metadata/source fields, not only body text.",
-		"PR #56 commit 2e4b0cc changed evalListEdges LIMIT handling for recall feedback export.",
-		[]string{"PR #56", "2e4b0cc", "evalListEdges"},
-		[]string{"PR #65", "2e4b0aa"},
-		[]evalharness.CorpusItem{
-			feedbackCorpusItem(requiredID, "Recall feedback export fix", "RAG accuracy note. The evalListEdges LIMIT handling was corrected for feedback export pagination.", "identifier_metadata", "required", 1.0, caseID, map[string]any{
-				"pr_number": 56,
-				"commit":    "2e4b0cc",
-				"function":  "evalListEdges",
-				"file":      "internal/tools/registry/evaluation_tools.go",
-			}),
-			feedbackCorpusItem(negative1ID, "Different recall feedback PR", "PR #65 adjusted unrelated evaluation pagination defaults and did not touch evalListEdges LIMIT behavior.", "identifier_metadata", "hard_negative", 0.2, caseID, map[string]any{
-				"pr_number": 65,
-				"commit":    "2e4b0aa",
-				"function":  "evalListFeedbackSummaries",
-			}),
-			feedbackCorpusItem(negative2ID, "Copied eval note", "A copied note mentions evalListEdges LIMIT handling, but it belongs to an abandoned draft and has no PR #56 fix.", "identifier_metadata", "hard_negative", 0.2, caseID, map[string]any{
-				"branch": "draft/feedback-export",
-			}),
-		},
-		nil,
-	)
-}
-
-func buildFeedbackToolIdentifierCase() generatedCase {
-	caseID := feedbackCaseID(2)
-	requiredID := caseID + "_required"
-	negative1ID := caseID + "_negative_01"
-	negative2ID := caseID + "_negative_02"
-	query := "Which recall feedback tool should be used for feedback:read?"
-	return feedbackGeneratedCase(
-		caseID,
-		"tool_identifier_metadata",
-		query,
-		"Find the tool identifier even when it appears only in labels and metadata.",
-		"Use eval_list_recall_feedback_events for feedback:read.",
-		[]string{"eval_list_recall_feedback_events", "feedback:read"},
-		[]string{"recall_memory", "dream:read"},
-		[]evalharness.CorpusItem{
-			feedbackCorpusItem(requiredID, "Recall feedback reader tool", "The recall feedback read operation should use the evaluation feedback event listing tool.", "tool_identifier_metadata", "required", 1.0, caseID, map[string]any{
-				"tool_name":  "eval_list_recall_feedback_events",
-				"capability": "feedback:read",
-			}),
-			feedbackCorpusItem(negative1ID, "General recall tool", "recall_memory retrieves user memories but does not list recall feedback rating events.", "tool_identifier_metadata", "hard_negative", 0.2, caseID, map[string]any{
-				"tool_name": "recall_memory",
-			}),
-			feedbackCorpusItem(negative2ID, "Dream feedback reader", "dream_list reads dream hypotheses and dream feedback, not recall feedback events.", "tool_identifier_metadata", "hard_negative", 0.2, caseID, map[string]any{
-				"capability": "dream:read",
-			}),
-		},
-		nil,
-	)
-}
-
-func buildFeedbackPathConfigCase() generatedCase {
-	caseID := feedbackCaseID(3)
-	requiredID := caseID + "_required"
-	negative1ID := caseID + "_negative_01"
-	negative2ID := caseID + "_negative_02"
-	query := "What limit did .textlintrc.json and scripts/ci-check.sh set for max file length?"
-	return feedbackGeneratedCase(
-		caseID,
-		"path_config_identifier",
-		query,
-		"Match exact config and script paths, not a generic lint rule.",
-		"The file length limit is 700 lines in .textlintrc.json and scripts/ci-check.sh.",
-		[]string{"700", ".textlintrc.json", "scripts/ci-check.sh"},
-		[]string{"500", ".eslintrc.json"},
-		[]evalharness.CorpusItem{
-			feedbackCorpusItem(requiredID, "Textlint file length gate", "CI policy sets the maximum file length gate to 700 lines for the text lint job.", "path_config_identifier", "required", 1.0, caseID, map[string]any{
-				"config_key": "max_file_length",
-				"config":     ".textlintrc.json",
-				"script":     "scripts/ci-check.sh",
-				"limit":      700,
-			}),
-			feedbackCorpusItem(negative1ID, "ESLint gate", "The JavaScript lint job in .eslintrc.json uses a 500 line warning and is unrelated to textlint.", "path_config_identifier", "hard_negative", 0.2, caseID, map[string]any{
-				"config": ".eslintrc.json",
-				"limit":  500,
-			}),
-			feedbackCorpusItem(negative2ID, "Old textlint draft", "An old textlint draft proposed 600 lines, but scripts/ci-check.sh never used that value.", "path_config_identifier", "hard_negative", 0.2, caseID, map[string]any{
-				"config": ".textlintrc.draft.json",
-				"limit":  600,
-			}),
-		},
-		nil,
-	)
-}
-
-func buildFeedbackAuthorityFactCase() generatedCase {
-	caseID := feedbackCaseID(4)
-	requiredID := caseID + "_required"
-	negative1ID := caseID + "_negative_01"
-	negative2ID := caseID + "_negative_02"
-	requiredFactID := typedFactSourceDocID(requiredID, 1)
-	negative1FactID := typedFactSourceDocID(negative1ID, 1)
-	negative2FactID := typedFactSourceDocID(negative2ID, 1)
-	query := "What is the source of truth for issue DM-412 recall feedback rating?"
-	corpus := []evalharness.CorpusItem{
-		feedbackCorpusItem(requiredID, "Authoritative recall feedback issue DM-412", "Authoritative source of truth. Issue DM-412 says low recall ratings are fixed by deterministic identifier extraction at write and recall time.", "authority_fact_identifier", "required", 1.0, caseID, map[string]any{
-			"issue_id": "DM-412",
-			"version":  "v1.6.4",
-		}, feedbackClaim("issue DM-412", "profile_fact", "low recall ratings require deterministic identifier extraction at write and recall time", map[string]any{
-			"issue_id": "DM-412",
-			"version":  "v1.6.4",
-		})),
-		feedbackCorpusItem(negative1ID, "Informal recall feedback issue DM-412", "Informal chat guessed that issue DM-412 should be fixed by asking the client LLM to include more context.", "authority_fact_identifier", "hard_negative", 0.2, caseID, map[string]any{
-			"issue_id": "DM-412",
-			"source":   "chat",
-		}, feedbackClaim("informal note DM-412", "profile_fact", "ask the client LLM to include more context", map[string]any{
-			"issue_id": "DM-412",
-		})),
-		feedbackCorpusItem(negative2ID, "Neighbor recall feedback issue DM-421", "Authoritative note for issue DM-421 covers UI copy changes, not identifier extraction.", "authority_fact_identifier", "hard_negative", 0.2, caseID, map[string]any{
-			"issue_id": "DM-421",
-		}, feedbackClaim("issue DM-421", "profile_fact", "UI copy changes", map[string]any{
-			"issue_id": "DM-421",
-		})),
-	}
-	return feedbackGeneratedCase(
-		caseID,
-		"authority_fact_identifier",
-		query,
-		"Prefer the active fact for the exact issue id over a matching raw fragment or informal note.",
-		"The source of truth for DM-412 is deterministic identifier extraction at write and recall time.",
-		[]string{"DM-412", "deterministic identifier extraction"},
-		[]string{"client LLM", "DM-421"},
-		corpus,
-		func(qrel *evalharness.QRel) {
-			qrel.RequiredRefs = []evalharness.Ref{{Type: "fact", SourceDocID: requiredFactID, Grade: 3, Reason: "authoritative issue fact"}}
-			qrel.BadRefs = []evalharness.Ref{
-				{Type: "fact", SourceDocID: negative1FactID, Grade: -1, Reason: "informal non-authoritative issue claim"},
-				{Type: "fact", SourceDocID: negative2FactID, Grade: -1, Reason: "neighbor issue fact"},
-			}
-			qrel.RequiredEvidenceRefs = []evalharness.Ref{{Type: "fragment", SourceDocID: requiredID, Grade: 1, Reason: "supporting authoritative issue source"}}
-			qrel.BadEvidenceRefs = []evalharness.Ref{
-				{Type: "fragment", SourceDocID: negative1ID, Grade: -1, Reason: "informal source"},
-				{Type: "fragment", SourceDocID: negative2ID, Grade: -1, Reason: "neighbor issue source"},
-			}
-		},
-	)
-}
-
-func buildFeedbackBranchRunCase() generatedCase {
-	caseID := feedbackCaseID(5)
-	requiredID := caseID + "_required"
-	negative1ID := caseID + "_negative_01"
-	negative2ID := caseID + "_negative_02"
-	query := "What result belongs to branch exp/recall-typed-expired-currentness run local_eval_1k_v2?"
-	return feedbackGeneratedCase(
-		caseID,
-		"branch_run_identifier",
-		query,
-		"Use exact branch and run identifiers so unrelated eval summaries do not win.",
-		"The exp/recall-typed-expired-currentness run local_eval_1k_v2 passed typed currentness hardcases.",
-		[]string{"exp/recall-typed-expired-currentness", "local_eval_1k_v2", "passed"},
-		[]string{"failed", "local_train_relational_100k_v1"},
-		[]evalharness.CorpusItem{
-			feedbackCorpusItem(requiredID, "Typed expired currentness run", "Evaluation summary: typed currentness hardcases passed for the target experiment.", "branch_run_identifier", "required", 1.0, caseID, map[string]any{
-				"branch": "exp/recall-typed-expired-currentness",
-				"run_id": "local_eval_1k_v2",
-				"status": "passed",
-			}),
-			feedbackCorpusItem(negative1ID, "Different branch run", "Evaluation summary: exp/recall-dream-rerank passed dream filtering checks for local_eval_1k_v2.", "branch_run_identifier", "hard_negative", 0.2, caseID, map[string]any{
-				"branch": "exp/recall-dream-rerank",
-				"run_id": "local_eval_1k_v2",
-			}),
-			feedbackCorpusItem(negative2ID, "Different dataset run", "Evaluation summary: exp/recall-typed-expired-currentness failed a training-only stress run.", "branch_run_identifier", "hard_negative", 0.2, caseID, map[string]any{
-				"branch": "exp/recall-typed-expired-currentness",
-				"run_id": "local_train_relational_100k_v1",
-				"status": "failed",
-			}),
-		},
-		nil,
-	)
-}
-
-func buildFeedbackDreamFilterCase() generatedCase {
-	caseID := feedbackCaseID(6)
-	requiredID := caseID + "_required"
-	negative1ID := caseID + "_negative_01"
-	negative2ID := caseID + "_negative_02"
-	requiredDreamID := caseID + "_expected_dream"
-	badDreamID := caseID + "_bad_dream"
-	query := "Which dream is related to feedback:read issue DM-512 recall filtering?"
-	corpus := []evalharness.CorpusItem{
-		feedbackCorpusItem(requiredID, "Dream anchor for feedback read issue", "Issue DM-512 describes feedback:read recall filtering and asks that related dreams be reranked with the same evidence anchors.", "dream_relevance", "required", 1.0, caseID, map[string]any{
-			"issue_id":   "DM-512",
-			"capability": "feedback:read",
-		}),
-		feedbackCorpusItem(negative1ID, "Unrelated dream anchor", "Issue DM-904 covers deployment calendar formatting and should not affect feedback:read recall filtering dreams.", "dream_relevance", "hard_negative", 0.2, caseID, map[string]any{
-			"issue_id": "DM-904",
-		}),
-		feedbackCorpusItem(negative2ID, "Generic dream policy", "Generic note: dreams are hypotheses and should not be treated as validated memory.", "dream_relevance", "hard_negative", 0.2, caseID, map[string]any{
-			"policy": "dreams",
-		}),
-	}
-	return feedbackGeneratedCase(
-		caseID,
-		"dream_relevance",
-		query,
-		"Return only the dream anchored to the recalled feedback:read issue and filter unrelated dreams.",
-		"The related dream is the DM-512 feedback:read reranking hypothesis.",
-		[]string{"DM-512", "feedback:read"},
-		[]string{"DM-904"},
-		corpus,
-		func(qrel *evalharness.QRel) {
-			qrel.RequiredDreamRefs = []evalharness.Ref{{Type: "dream", SourceDocID: requiredDreamID, Grade: 3, Reason: "dream anchored to required recall hit"}}
-			qrel.BadDreamRefs = []evalharness.Ref{{Type: "dream", SourceDocID: badDreamID, Grade: -1, Reason: "dream anchored to unrelated issue"}}
-		},
-		func(gc *generatedCase) {
-			gc.Case.IncludeDreams = true
-			gc.Dreams = []evalharness.ExpectedDream{
-				{
-					SourceDocID: requiredDreamID,
-					CaseID:      caseID,
-					Hypothesis:  "feedback:read issue DM-512 may require dream reranking against the same recall evidence anchors.",
-					SourceRefs: []evalharness.Ref{
-						{Type: "fragment", SourceDocID: requiredID},
-						{Type: "fragment", SourceDocID: negative2ID},
-					},
-					Labels: []string{"eval", "dream_relevance", "expected_dream"},
-					Metadata: map[string]any{
-						"case_id":    caseID,
-						"issue_id":   "DM-512",
-						"capability": "feedback:read",
-					},
-				},
-				{
-					SourceDocID: badDreamID,
-					CaseID:      caseID,
-					Hypothesis:  "DM-904 deployment calendar formatting may need a separate dream.",
-					SourceRefs: []evalharness.Ref{
-						{Type: "fragment", SourceDocID: negative1ID},
-						{Type: "fragment", SourceDocID: negative2ID},
-					},
-					Bad:    true,
-					Labels: []string{"eval", "dream_relevance", "bad_dream"},
-					Metadata: map[string]any{
-						"case_id":  caseID,
-						"issue_id": "DM-904",
-					},
-				},
-			}
-		},
-	)
-}
-
-func feedbackGeneratedCase(
-	caseID, category, query, expected, answer string,
-	mustInclude, mustNotInclude []string,
-	corpus []evalharness.CorpusItem,
-	qrelMutator func(*evalharness.QRel),
-	extraMutators ...func(*generatedCase),
-) generatedCase {
-	requiredID := caseID + "_required"
-	negative1ID := caseID + "_negative_01"
-	negative2ID := caseID + "_negative_02"
-	slices := []string{"feedback_regression", "adversarial", category}
-	qrel := evalharness.QRel{
-		CaseID: caseID,
-		RequiredRefs: []evalharness.Ref{
-			{Type: "fragment", SourceDocID: requiredID, Grade: 3, Reason: "required feedback hardcase record"},
-		},
-		BadRefs: []evalharness.Ref{
-			{Type: "fragment", SourceDocID: negative1ID, Grade: -1, Reason: "feedback hardcase decoy"},
-			{Type: "fragment", SourceDocID: negative2ID, Grade: -1, Reason: "feedback hardcase decoy"},
-		},
-		RequiredEvidenceRefs: []evalharness.Ref{
-			{Type: "fragment", SourceDocID: requiredID, Grade: 1, Reason: "required source evidence"},
-		},
-		BadEvidenceRefs: []evalharness.Ref{
-			{Type: "fragment", SourceDocID: negative1ID, Grade: -1, Reason: "misleading source evidence"},
-			{Type: "fragment", SourceDocID: negative2ID, Grade: -1, Reason: "misleading source evidence"},
-		},
-	}
-	if qrelMutator != nil {
-		qrelMutator(&qrel)
-	}
-	gc := generatedCase{
-		Case: evalharness.Case{
-			CaseID:           caseID,
-			Query:            query,
-			TaskType:         "memory_recall",
-			Difficulty:       "hard",
-			Slices:           slices,
-			ExpectedBehavior: expected,
-			KnownAt:          "2026-07-08T00:00:00Z",
-			Limit:            5,
-		},
-		SuiteCase: evalharness.SuiteCase{CaseID: caseID, Weight: 1, Slices: slices},
-		Corpus:    corpus,
-		QRel:      qrel,
-		Answer: evalharness.AnswerLabel{
-			CaseID:             caseID,
-			ReferenceAnswer:    answer,
-			MustInclude:        mustInclude,
-			MustNotInclude:     mustNotInclude,
-			ExpectedBehavior:   expected,
-			GroundednessPolicy: "Use the exact feedback hardcase memory; reject close identifier, source, and dream decoys.",
-		},
-		HardNegatives: []hardNegativeRow{
-			{CaseID: caseID, Category: category, SourceDocID: negative1ID, Reason: "Close lexical overlap but wrong identifier, authority, or anchor."},
-			{CaseID: caseID, Category: category, SourceDocID: negative2ID, Reason: "Neighbor or stale feedback item that should not answer the query."},
-		},
-		Transform: transformRow{CaseID: caseID, Category: category, Transform: expected, DocsPerCase: len(corpus), RequiredDocID: requiredID},
-	}
-	for _, mutate := range extraMutators {
-		mutate(&gc)
-	}
-	return gc
-}
-
-func feedbackCorpusItem(sourceDocID, title, content, category, role string, quality float64, caseID string, metadata map[string]any, claims ...evalharness.TypedClaim) evalharness.CorpusItem {
-	itemMetadata := map[string]any{
-		"case_id":  caseID,
-		"role":     role,
-		"category": category,
-	}
-	for key, value := range metadata {
-		itemMetadata[key] = value
-	}
-	for i := range claims {
-		if claims[i].IdempotencyKey == "" {
-			claims[i].IdempotencyKey = fmt.Sprintf("eval:%s:claim:%d", sourceDocID, i+1)
-		}
-		if claims[i].ExtractConf == 0 {
-			claims[i].ExtractConf = 0.98
-		}
-		if claims[i].ResolutionConf == 0 {
-			claims[i].ResolutionConf = 0.98
-		}
-		if claims[i].ExtractionModel == "" {
-			claims[i].ExtractionModel = "dense-mem.eval-feedback-seedgen"
-		}
-		if claims[i].ExtractionVersion == "" {
-			claims[i].ExtractionVersion = "v1"
-		}
-		if claims[i].PipelineRunID == "" {
-			claims[i].PipelineRunID = "eval:" + sourceDocID
-		}
-		if claims[i].Classification == nil {
-			claims[i].Classification = map[string]any{}
-		}
-		claims[i].Classification["case_id"] = caseID
-		claims[i].Classification["category"] = category
-		claims[i].Classification["role"] = role
-		claims[i].Classification["confidentiality"] = "public"
-		claims[i].Classification["pii"] = "synthetic"
-	}
-	return evalharness.CorpusItem{
-		SourceDocID:   sourceDocID,
-		Title:         title,
-		Content:       content,
-		SourceDataset: presetFeedbackRecallHardcases,
-		SourceType:    "synthetic",
-		Authority:     role,
-		SourceQuality: quality,
-		Labels:        []string{"eval", "feedback_regression", category, role},
-		Metadata:      itemMetadata,
-		Claims:        claims,
-		AutoPromote:   len(claims) > 0,
-	}
-}
-
-func feedbackClaim(subject, predicate, object string, classification map[string]any) evalharness.TypedClaim {
-	claimClassification := map[string]any{
-		"confidentiality": "public",
-		"pii":             "synthetic",
-	}
-	for key, value := range classification {
-		claimClassification[key] = value
-	}
-	return evalharness.TypedClaim{
-		Subject:        subject,
-		Predicate:      predicate,
-		Object:         object,
-		Modality:       "assertion",
-		Polarity:       "+",
-		ExtractConf:    0.98,
-		ResolutionConf: 0.98,
-		Classification: claimClassification,
-	}
-}
-
-func feedbackCaseID(index int) string {
-	return fmt.Sprintf("%s_%03d", presetFeedbackRecallHardcases, index)
-}
-
 func buildLocalEval1KV2() []generatedCase {
 	out := make([]generatedCase, 0, 1000)
 	for i := 1; i <= 100; i++ {
@@ -1466,16 +968,6 @@ This seed is synthetic project-local evaluation data generated by cmd/eval-seedg
 
 It contains no third-party source text. Use it for local Dense-Mem recall quality
 tests and regression gates within this repository.
-`) + "\n"
-}
-
-func feedbackHardcaseLicense() string {
-	return strings.TrimSpace(`# feedback_recall_hardcases_v1 Licenses
-
-This seed is synthetic project-local recall feedback regression data generated by cmd/eval-seedgen.
-
-It contains no production data and no third-party source text. Use it for local
-Dense-Mem recall, evidence, identifier, and dream filtering regression tests.
 `) + "\n"
 }
