+++ internal/service/dreamservice/service.go
@@ -13,7 +13,6 @@ import (
 	"github.com/markhuangai/dense-mem/internal/domain"
 	"github.com/markhuangai/dense-mem/internal/fulltextquery"
 	"github.com/markhuangai/dense-mem/internal/observability"
-	"github.com/markhuangai/dense-mem/internal/recallident"
 	"github.com/markhuangai/dense-mem/internal/service/memoryservice"
 	neo4jstorage "github.com/markhuangai/dense-mem/internal/storage/neo4j"
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
@@ -903,24 +902,6 @@ func dreamContentHash(d *domain.Dream) string {
 		string(data),
 	}, "\x00")))
 	return hex.EncodeToString(h[:])
-}
-
-func dreamRecallFields(d *domain.Dream) (string, []string) {
-	if d == nil {
-		return "", nil
-	}
-	parts := []string{
-		d.Hypothesis,
-		d.WhatIf,
-		d.PossibleOutcome,
-		d.Rationale,
-		d.CycleRunID,
-		d.GeneratorModel,
-	}
-	for _, ref := range d.SourceRefs {
-		parts = append(parts, ref.Type, ref.ID)
-	}
-	return recallident.BuildRecallText(parts...)
 }
 
 func localRunDate(now time.Time, cfg EffectiveConfig) string {
